+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Version 1.24 (not yet released)
 ------------
 
+* Added more fields to contribution creation (fee amount, check number, receipt date, etc.)
+
 Version 1.23
 ------------