+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version 1.24 (not yet released)
 
 * Added more fields to contribution creation (fee amount, check number, receipt date, etc.)
 * Added action Get Country By Id.
+* Make Activity Date optional - defaults to 'now'
 
 Version 1.23
 ------------