+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 1.24 (not yet released)
 * Added more fields to contribution creation (fee amount, check number, receipt date, etc.)
 * Added action Get Country By Id.
 * Make Activity Date optional - defaults to 'now'
+* Fixed issue with country custom fields.
 
 Version 1.23
 ------------